# Speed-up generate cluster image