--- a/napari_clusters_plotter/_utilities.py
+++ b/napari_clusters_plotter/_utilities.py
@@ -299,8 +299,13 @@ def generate_cluster_image(label_image, label_list, predictionlist):
     """
 
     # label_list can be removed from method.
-
     predictionlist_new = np.array(predictionlist) + 1
+    plist = np.zeros(np.max(label_image)+1,dtype=np.uint32)
+    plist[label_list] = predictionlist_new
+
+    predictionlist_new = plist
+
+
 
     return predictionlist_new[label_image].astype(
         "uint32"