The generate_cluster_image method is the most computationally intensive part when working with large data like mine (32 million data points). As already written in #282, the total time from circling a cluster to visualizing the highlighted cluster is 17 seconds for my data.

This PR reduces it by another 6 seconds, and thus together with PR #282 from 17 seconds to 6 seconds - which is a much better user experience.

The PR works beautifully on my data, but I'm a bit skeptical that it will work with other workflows. So it would require some testing on your part :-) 

Please let me know your thoughts!

Best,
Thorsten