@thorstenwagner 2023-12-07
Alright! Now all tests pass and its still faster :-) What do you think?