@thorstenwagner 2023-12-08
Here are a script to evaluate the speedup:
```python
import numpy as np
from skimage.util import map_array
from time import perf_counter

shape = (1,1024,1024)
total = shape[0]*shape[1]*shape[2]
NUM_LABELS=100000
input_data = np.random.randint(NUM_LABELS,size=total).reshape(shape).astype("int64")
from_values = np.arange(NUM_LABELS)
to_values = np.copy(from_values)[::-1]

def generate_cluster_image(label_image, label_list, predictionlist):
    """
    Generates a clusters image from a label image and a list of cluster predictions,
    where each label value corresponds to the cluster identity.
    It is assumed that len(predictionlist) == max(label_image)

    Parameters
    ----------
    label_image: ndarray or dask array
        Label image used for cluster predictions
    predictionlist: Array-like
        An array containing cluster identities for each label

    Returns
    ----------
    ndarray: The clusters image as a numpy array.
    """

    # label_list can be removed from method.
    predictionlist_new = np.array(predictionlist) + 1
    plist = np.zeros(np.max(label_image) + 1, dtype=np.uint32)
    plist[label_list] = predictionlist_new

    predictionlist_new = plist

    return predictionlist_new[label_image]

def generate_cluster_image_old(label_image, label_list, predictionlist):
    """
    Generates a clusters image from a label image and a list of cluster predictions,
    where each label value corresponds to the cluster identity.
    It is assumed that len(predictionlist) == max(label_image)

    Parameters
    ----------
    label_image: ndarray or dask array
        Label image used for cluster predictions
    predictionlist: Array-like
        An array containing cluster identities for each label

    Returns
    ----------
    ndarray: The clusters image as a numpy array.
    """
    from skimage.util import map_array

    # reforming the prediction list, this is done to account
    # for cluster labels that start at 0, conveniently hdbscan
    # labelling starts at -1 for noise, removing these from the labels
    predictionlist_new = np.array(predictionlist) + 1
    label_list = np.array(label_list)

    return map_array(np.asarray(label_image), label_list, predictionlist_new).astype(
        "uint32"
    )

t1 = perf_counter()
res_new = generate_cluster_image(input, from_values, to_values)
t_new = perf_counter() - t1
print(t_new)

t1 = perf_counter()
res_old = generate_cluster_image_old(input, from_values, to_values)
t_old = perf_counter() - t1
print(t_old)

print(f"Speedup {t_old/t_new}")

print(np.array_equal(res_new,res_old))
```

The new function is like 30x-35x faster.