@Cryaaa 2023-12-11
Hey @thorstenwagner!

> Thanks for the merge! One question: what do you want to happen if from_values and to_values are only a subset of the labels from label image?
> 
> Right now they are mapped to zero. But we can easily modify it so that they dont get changed?

I'm not sure what benefit that would have exactly? The cluster image is just supposed to show the data points that have been clustered in some way. It might also be a fringe case but if you just transfer the original label and the original label coincides with the cluster label you have misclassified objects. Other than that if you are interested in the unclassified labels you could just extract them from the table and process them separately.